// src/generated/models/Aur_consequencesesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Aur_consequencesesaur_action = {
  1: 'Decline',
  2: 'RefertoUnderwriter',
  3: 'EscalatetoHeadofAviation',
  4: 'PropertyorReinsuranceTeam',
  5: 'RequestMissingInformation',
  6: 'Document',
  7: 'GenerateDeclineLetter',
  8: 'GenerateNoticeofCancellation',
  9: 'UpdateRiskScore',
  751820001: 'UpdateRiskSummary',
  10: 'Email'
} as const;
export type Aur_consequencesesaur_action = keyof typeof Aur_consequencesesaur_action;
export const Aur_consequencesesaur_type = {
  1: 'CaseControl',
  2: 'Notification',
  3: 'Document',
  4: 'Risk',
  5: 'Email'
} as const;
export type Aur_consequencesesaur_type = keyof typeof Aur_consequencesesaur_type;
export const Aur_consequencesesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Aur_consequencesesstatecode = keyof typeof Aur_consequencesesstatecode;
export const Aur_consequencesesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Aur_consequencesesstatuscode = keyof typeof Aur_consequencesesstatuscode;

export interface Aur_consequencesesBase {
  aur_action?: Aur_consequencesesaur_action;
  "aur_business_rule@odata.bind"?: string;
  aur_consequencesid: string;
  "aur_DocumentTemplate@odata.bind"?: string;
  aur_name: string;
  aur_notification_text?: string;
  aur_type?: Aur_consequencesesaur_type;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Aur_consequencesesstatecode;
  statuscode?: Aur_consequencesesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Aur_consequenceses extends Aur_consequencesesBase {
  aur_actionname?: string;
  aur_business_rulename?: string;
  aur_documenttemplatename?: string;
  aur_typename?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  aur_business_rule?: object;
  _aur_business_rule_value?: string;
  aur_documenttemplate?: object;
  _aur_documenttemplate_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}